TASK - 49179 :Import notes from a zip file

+++ notes-service/src/main/java/org/exoplatform/wiki/service/impl/NoteServiceImpl.java
@@ -1444,7 +1444,9 @@ public class NoteServiceImpl implements NoteService {
   }
 
   public static void cleanUp(File file) throws IOException {
-    Files.delete(file.toPath());
+    if(Files.exists(file.toPath())){
+      Files.delete(file.toPath());
+    }
   }
 
 }
